refactor: GameCount 원시값 포장 및 테스트 수정함
